requirements.txt
# Code analysis
flake8
flake8-docstrings
flake8-polyfill
pylint

# Database
mysql-connector-python
